reusable table tabs for class and project
